Move driving pause states into their own module
